+++ src/main/java/com/zeylin/kafkademotemplates/KafkaDemoTemplatesApplication.java
@@ -1,9 +1,11 @@
-package com.example.kafkademotemplates;
+package com.zeylin.kafkademotemplates;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.kafka.annotation.EnableKafka;
 
 @SpringBootApplication
+@EnableKafka
 public class KafkaDemoTemplatesApplication {
 
 	public static void main(String[] args) {