+++ src/main/java/com/zeylin/kafkademotemplates/config/KafkaConsumerConfig.java
@@ -1,0 +1,57 @@
+package com.zeylin.kafkademotemplates.config;
+
+import com.zeylin.kafkademotemplates.config.serializers.SpaceshipDeserializer;
+import com.zeylin.kafkademotemplates.dto.SpaceshipDto;
+import com.zeylin.kafkademotemplates.service.SpaceshipListener;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.common.serialization.LongDeserializer;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.config.KafkaListenerContainerFactory;
+import org.springframework.kafka.core.ConsumerFactory;
+import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+@Slf4j
+public class KafkaConsumerConfig {
+
+    @Value("${kafka.server}")
+    private String kafkaServer;
+
+    @Value("${kafka.group}")
+    private String kafkaGroupId;
+
+    @Bean
+    public KafkaListenerContainerFactory<?> kafkaListenerContainerFactory() {
+        ConcurrentKafkaListenerContainerFactory<Long, SpaceshipDto> factory =
+                new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(consumerFactory());
+        factory.setErrorHandler((e, consumerRecord) ->
+                log.warn("ERROR: Inbound message from topic: {}, message: {}", SpaceshipListener.TOPIC, e.getCause())
+        );
+        return factory;
+    }
+
+    @Bean
+    public ConsumerFactory<Long, SpaceshipDto> consumerFactory() {
+        return new DefaultKafkaConsumerFactory<>(consumerConfigs());
+    }
+
+    @Bean
+    public Map<String, Object> consumerConfigs() {
+        Map<String, Object> props = new HashMap<>();
+        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaServer);
+        props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, LongDeserializer.class);
+        props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, SpaceshipDeserializer.class);
+        props.put(ConsumerConfig.GROUP_ID_CONFIG, kafkaGroupId);
+        props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, true);
+        return props;
+    }
+
+}