+++ src/main/java/com/zeylin/kafkademotemplates/config/KafkaProducerConfig.java
@@ -1,0 +1,49 @@
+package com.zeylin.kafkademotemplates.config;
+
+import com.zeylin.kafkademotemplates.config.serializers.SpaceshipSerializer;
+import com.zeylin.kafkademotemplates.dto.SpaceshipDto;
+import org.apache.kafka.clients.producer.ProducerConfig;
+import org.apache.kafka.common.serialization.LongSerializer;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.DefaultKafkaProducerFactory;
+import org.springframework.kafka.core.KafkaTemplate;
+import org.springframework.kafka.core.ProducerFactory;
+import org.springframework.kafka.support.converter.StringJsonMessageConverter;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Configuration
+public class KafkaProducerConfig {
+
+    @Value("${kafka.server}")
+    private String kafkaServer;
+
+    @Value("${kafka.producer.id}")
+    private String kafkaProducerId;
+
+    @Bean
+    public Map<String, Object> producerConfigs() {
+        Map<String, Object> props = new HashMap<>();
+        props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaServer);
+        props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
+        props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, SpaceshipSerializer.class);
+        props.put(ProducerConfig.CLIENT_ID_CONFIG, kafkaProducerId);
+        return props;
+    }
+
+    @Bean
+    public ProducerFactory<Long, SpaceshipDto> producerStarshipFactory() {
+        return new DefaultKafkaProducerFactory<>(producerConfigs());
+    }
+
+    @Bean
+    public KafkaTemplate<Long, SpaceshipDto> kafkaTemplate() {
+        KafkaTemplate<Long, SpaceshipDto> template = new KafkaTemplate<>(producerStarshipFactory());
+        template.setMessageConverter(new StringJsonMessageConverter());
+        return template;
+    }
+
+}