+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0-SNAPSHOT] - 2019-12-27
+### Added
+- Send and receive messages with Kafka
+- Kafka producer and consumer configs
+- Readme guide
+
 ## [1.0.0-SNAPSHOT] - 2019-12-27
 ### Added
 - Initial commit